requirements.txt
flask
python-dotenv
google-generativeai
youtube-transcript-api